Cleaned up Duration for the standard library.
 - Removed Timeout from the library. Each library should provide its own implementation of Timeout
 - Removed duration package object and replaced it with regular object
 - Removed usages of post Java 1.5 TimeUnit members
 - Added factory methods for FiniteDuration
 - Cleaned up some unnecessary comments

Review by @phaller
